+++ editor/CHANGES.md
@@ -12,6 +12,9 @@
 - **Fixed**: Unable to change Z index and scale when using text entities with auto set origin to center enabled
 - **Fixed**: Undoing entity delete could place it at wrong position
 - **Fixed**: Crash when user tried to drag atlas region before any scene had been opened 
+- **Fixed**: Twitter view URL parsing issues
+- **Fixed**: When opening new scene selection tool as used instead of currently active tool
+- **Fixed**: [https://github.com/kotcrab/vis-editor/issues/181] - invalid rotate and scale tools size when using non default pixels per unit values 
 
 #### Version: 0.3.1
 - **Added**: Scale entities tool